// src/hooks/api/queryKeys.ts
export enum QUERY_KEY {
  MY_ACCOUNT = 'MY_ACCOUNT',
  USER = 'USER',
  USERS = 'USERS',
  ROOMS = 'ROOMS',
  ROOM_SUMMARY = 'ROOM_SUMMARY',
  ROOMS_FLOOR_PLAN = 'ROOMS_FLOOR_PLAN',
  AVAILABLE_ROOMS = 'AVAILABLE_ROOMS',
  ROOM = 'ROOM',
  BOOKINGS = 'BOOKINGS',
  BOOKING = 'BOOKING',
  FINANCE_SUMMARY_THIS_MONTH = 'FINANCE_SUMMARY_THIS_MONTH',
}